Reduce the SSL buffer size from 16MiB to 16KiB
Unfortunately there are lots of hidden datacopies in the datapath
which will kill your performance with large buffer sizes:

* Lwt_ssl.read_bytes creates a temporary *string* of the same size
  as your Bigarray. It calls SSL.read and then blits the string into
  your Bigarray.

* SSL.read == ocaml_ssl_read which creates another temporary C buffer
  using malloc of the same size as Lwt_ssl's string. It will then call
  SSL_read, which typically returns 16KiB. It will then memmove()
  the data back into Lwt_ssl's string.

Therefore with a requested 16MiB buffer size, every 16KiB of data read
involves immediately throwing away 16MiB of string in the OCaml heap
and 16MiB of string in the C heap, despite the fact the application
is being careful to use Bigarray-based APIs and large buffers for
efficiency. The unhappy result is that transfers slow to a crawl
as the process becomes GC limited.

I hope this never happens to you.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>